UAT products test

+++ lib/actions/stock.actions.ts
@@ -17,7 +17,8 @@ import { redirect } from 'next/navigation'
 const {
     APPWRITE_DATABASE: DATABASE_ID,
     INVENTORY_VARIANTS_COLLECTION: INVENTORY_COLLECTION_ID,
-    STOCK_COLLECTION: STOCK_COLLECTION_ID
+    STOCK_COLLECTION: STOCK_COLLECTION_ID,
+    ALARM_QUANTITY = 3
 } = process.env;
 
 const checkRequirements = async (collectionId: string | undefined) => {
@@ -39,34 +40,75 @@ const checkRequirements = async (collectionId: string | undefined) => {
 
 export const createItem = async (items: Stock[]) => {
   const { database, businessId } = await checkRequirements(STOCK_COLLECTION_ID)
+  const alarmQuantity = ALARM_QUANTITY || 10;
+
   try {
-      for (const item of items) {
-        //increase item quantity
-        if ( item.item.actualQuantity < 0 ){
-          // this means usage had previously exceeded items in stock, so reduce the negative first
-          item.item.actualQuantity =  item.item.actualQuantity + item.quantity;
-        }else{
-          // Increase both quantity and actual quantity
-          item.item.actualQuantity =  item.item.actualQuantity + item.quantity;
-          item.item.quantity =  item.item.quantity + item.quantity;
-        }
+    for (const item of items) {
+      // Ensure quantities and values are parsed correctly and are numbers
+      const newQuantity = item.quantity || 0;
+      const newValue = item.value || 0;
 
-        //If actual quantity is still less than zero after the update, then leave quantity as is
-        if ( item.item.actualQuantity < 0 ){
-          item.item.quantity =  item.item.quantity;
-        }else{
-          // set item quantity to the actual quantity after the reduction of negative items
-          item.item.quantity =  item.item.actualQuantity;
-        }
+      const currentQuantity = parseInt(item.item.quantity) || 0;
+      const currentActualQuantity = parseInt(item.item.actualQuantity) || 0;
 
-        //Update stock status
-        if ( item.item.quantity === 0 ) {
+      const currentValue = parseFloat(item.item.value) || 0;
+      const currentActualValue = parseFloat(item.item.actualValue) || 0;
+
+      // Increase item quantity
+      if (currentActualQuantity < 0) {
+          // This means usage had previously exceeded items in stock, so reduce the negative first
+          item.item.actualQuantity = currentActualQuantity + newQuantity;
+          if (item.item.actualQuantity >= 0) {
+              item.item.quantity = item.item.actualQuantity;
+          } else {
+              item.item.quantity = 0;
+          }
+      } else {
+          // Increase quantity, value, and actual quantity
+          item.item.actualQuantity = currentActualQuantity + newQuantity;
+          item.item.quantity = currentQuantity + newQuantity;
+      }
+
+      // Update value
+      if (currentActualValue < 0) {
+          // Determine value per item and reduce negative value first
+          const valuePerItem = newValue / newQuantity;
+          item.item.actualValue = currentActualValue + newValue;
+
+          if (item.item.actualValue >= 0) {
+              item.item.value = item.item.actualValue;
+          } else {
+              item.item.value = 0;
+          }
+      } else {
+          item.item.actualValue = currentActualValue + newValue;
+          item.item.value = currentValue + newValue;
+      }
+
+      // If actual quantity is still less than zero after the update, then leave quantity as is
+      if (item.item.actualQuantity < 0) {
+          item.item.quantity = 0;
+      } else {
+          item.item.quantity = item.item.actualQuantity;
+      }
+
+      // If actual value is still less than zero after the update, then leave value as is
+      if (item.item.actualValue < 0) {
+          item.item.value = 0;
+      } else {
+          item.item.value = item.item.actualValue;
+      }
+
+      // Update stock status
+      if (item.item.quantity === 0) {
           item.item.status = InventoryStatus.OUT_OF_STOCK;
-        } else if (item.item.quantity <= item.item.lowQuantity) {
+      } else if (item.item.quantity <= item.item.lowQuantity) {
           item.item.status = InventoryStatus.LOW_STOCK;
-        } else {
+      } else if (item.item.quantity <= item.item.lowQuantity + alarmQuantity) {
+          item.item.status = InventoryStatus.ALARM;
+      } else {
           item.item.status = InventoryStatus.IN_STOCK;
-        }
+      }
 
         await database.createDocument(
           DATABASE_ID!,
@@ -77,7 +119,7 @@ export const createItem = async (items: Stock[]) => {
             businessId: businessId,
           }
         )
-      }
+    }
    } catch (error: any) {
       let errorMessage = 'Something went wrong with your request, please try again later.';
       if (error instanceof AppwriteException) {
@@ -100,8 +142,8 @@ export const list = async ( ) => {
 
     try {
       const items = await database.listDocuments(
-        DATABASE_ID,
-        STOCK_COLLECTION_ID,
+        DATABASE_ID!,
+        STOCK_COLLECTION_ID!,
         [Query.equal('businessId', businessId)],
       );
 
@@ -132,8 +174,6 @@ export const getItems = async (
     const { database, businessId } = await checkRequirements(STOCK_COLLECTION_ID)
 
     try {
-      const { database } = await createAdminClient();
-
       const queries = [];
       queries.push(Query.equal('businessId', businessId));
       queries.push(Query.orderDesc("$createdAt"));
@@ -144,7 +184,7 @@ export const getItems = async (
       }
 
       if (q) {
-        queries.push(Query.search('name', q));
+        //queries.push(Query.search('item.name', q));
       }
 
       if (status) {
@@ -152,8 +192,8 @@ export const getItems = async (
       }
 
       const items = await database.listDocuments(
-        DATABASE_ID,
-        STOCK_COLLECTION_ID,
+        DATABASE_ID!,
+        STOCK_COLLECTION_ID!,
         queries
       );
 
@@ -171,6 +211,7 @@ export const getItems = async (
       if(env == "development"){ console.error(error); }
 
       Sentry.captureException(error);
+
       throw Error(errorMessage);
     }
   }
@@ -182,8 +223,10 @@ export const getItems = async (
         const item = await database.listDocuments(
           DATABASE_ID!,
           STOCK_COLLECTION_ID!,
-          [Query.equal('$id', id)],
-          [Query.equal('businessId', businessId)],
+          [
+            Query.equal('$id', id),
+            Query.equal('businessId', businessId)
+          ]
         )
 
         if ( item.total < 1 ) return null;
@@ -203,7 +246,8 @@ export const getItems = async (
   }
 
 export const deleteItem = async ({ $id }: Stock) => {
-    const { database, businessId } = await checkRequirements(STOCK_COLLECTION_ID)
+    if (!$id) return null;
+    const { database } = await checkRequirements(STOCK_COLLECTION_ID)
 
     try {
       const item = await database.deleteDocument(
@@ -226,10 +270,11 @@ export const deleteItem = async ({ $id }: Stock) => {
 }
 
 export const updateItem = async (id: string, data: Stock) => {
-  const { database, businessId } = await checkRequirements(STOCK_COLLECTION_ID)
+  if (!id || !data) return null;
+  const { database } = await checkRequirements(STOCK_COLLECTION_ID)
 
   try {
-    const item = await database.updateDocument(
+    await database.updateDocument(
       DATABASE_ID!,
       STOCK_COLLECTION_ID!,
       id,
